+++ src/main/java/com/example/floud/dto/request/memoir/MemoirUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.floud.dto.request;
+package com.example.floud.dto.request.memoir;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
