+++ src/main/java/com/example/floud/dto/response/like/LikeMemoirListResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.floud.dto.response;
+package com.example.floud.dto.response.like;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
